+++ common/src/main/java/com/yanny/ali/registries/TradeLootCategory.java
@@ -7,11 +7,11 @@ import net.minecraft.world.item.ItemStack;
 import java.util.List;
 import java.util.regex.Pattern;
 
-public class GameplayLootCategory extends LootCategory<String> {
+public class TradeLootCategory extends LootCategory<String> {
     private final List<Pattern> patterns;
 
-    public GameplayLootCategory(String key, ItemStack icon, List<Pattern> patterns) {
-        super(key, icon, Type.GAMEPLAY);
+    public TradeLootCategory(String key, ItemStack icon, List<Pattern> patterns) {
+        super(key, icon, Type.TRADE);
         this.patterns = patterns;
     }